% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test2means.R
\name{auto_test2means}
\alias{auto_test2means}
\title{Compare two sample means or medians}
\usage{
auto_test2means(x, y)
}
\arguments{
\item{x, }{a vector of sample x}

\item{y, }{a vector of the second sample y}
}
\value{
the test output object
}
\description{
This is a function compare the two means or medians of
two samples based on the distributions of the samples.
}
